Improved new post/page template options.
1. Now users can set in _octopress.yml

```
new_post_template: post
new_page_template: page
```

Now when a user runs `octopress new post`, the template for a new post
will be read in from `[source]/_templates/post`. Of course you can have
a file extension if you like, but no file extension is
necessary since there is a default file extension for posts and pages
and that will be used regardless of what the template file is.

2. Now the template will only parse the YAML front-matter in Liquid.

This means that users can add any content they like after the yaml front
matter and it will not be parsed until it is built with Jekyll. This
means that liquid variables, tags will be preserved.